Add synchronization on AdminScopeData and put last_modified attribute

+++ src/main/java/com/sw/protection/backend/dao/impl/AdminDAOImpl.java
@@ -6,6 +6,7 @@
 
 package com.sw.protection.backend.dao.impl;
 
+import java.util.Date;
 import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -19,10 +20,12 @@ import org.hibernate.Transaction;
 import com.hazelcast.core.Hazelcast;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.IMap;
+import com.sw.protection.backend.common.Formatters;
 import com.sw.protection.backend.config.HibernateUtil;
 import com.sw.protection.backend.config.SharedInMemoryData;
 import com.sw.protection.backend.dao.AdminDAO;
 import com.sw.protection.backend.entity.Admin;
+import com.sw.protection.backend.entity.AdminScope;
 
 /**
  * Admin operation implementation
@@ -41,8 +44,9 @@ public class AdminDAOImpl implements AdminDAO {
     @Override
     public Admin getAdmin(String userName) {
 	session = HibernateUtil.getSessionFactory().getCurrentSession();
-	Transaction tr = session.beginTransaction();
+	Transaction tr = null;
 	try {
+	    tr = session.beginTransaction();
 	    List<Admin> adminAll = session.getNamedQuery(Admin.Constants.NAME_QUERY_FIND_BY_USER_NAME)
 		    .setParameter(Admin.Constants.PARAM_USER_NAME, userName).list();
 	    tr.commit();
@@ -69,32 +73,42 @@ public class AdminDAOImpl implements AdminDAO {
 
     @Override
     public void updateAdmin(Admin admin) {
-	session = HibernateUtil.getSessionFactory().getCurrentSession();
-	Transaction tr = session.beginTransaction();
+	Transaction tr = null;
 	try {
-	    // LOCKS.putIfAbsent(admin.getId(), new ReentrantLock());
-	    // LOCKS.get(admin.getId()).lock(); // lock by Admin ID
-	    // LOCK.putIfAbsent(admin.getId(), new ReentrantLock());
+	    // Lock by admin ID
 	    LOCK_MAP.lock(admin.getId());
 	    if (log.isDebugEnabled()) {
 		log.debug("Locked update operation by Admin ID " + admin.getId());
 	    }
-	    session.merge(admin);
-	    tr.commit();
 
-	    if (log.isDebugEnabled()) {
-		log.debug("Update Admin" + admin.toString());
+	    // Assume that the username never changed
+	    // check last modification
+	    if (admin.getLast_modified().equals(this.getAdmin(admin.getUser_name()).getLast_modified())) {
+		session = HibernateUtil.getSessionFactory().getCurrentSession();
+		tr = session.beginTransaction();
+		admin.setLast_modified(Formatters.formatDate(new Date()));
+		session.merge(admin);
+		tr.commit();
+
+		if (log.isDebugEnabled()) {
+		    log.debug("Update Admin" + admin.toString());
+		}
+	    } else {
+		if (log.isDebugEnabled()) {
+		    log.debug("This is not the latest modification of admin " + admin.toString() + " so cannot update");
+		}
+		// TODO:Create Exception
 	    }
 
 	} catch (RuntimeException ex) {
-	    tr.rollback();
 	    log.error(ex);
+	    tr.rollback();
 	    // TODO: capture the exception
 	} finally {
-	    // LOCKS.get(admin.getId()).unlock(); // Release by Admin ID
 	    if (log.isDebugEnabled()) {
 		log.debug("Releasing LOCK by Admin ID " + admin.getId());
 	    }
+	    // Unlock the lock by admin ID
 	    LOCK_MAP.unlock(admin.getId());
 	    // TODO: throw the captured exception
 	}
@@ -102,13 +116,22 @@ public class AdminDAOImpl implements AdminDAO {
 
     @Override
     public void deleteAdmin(Admin admin) {
-	session = HibernateUtil.getSessionFactory().getCurrentSession();
-	Transaction tr = session.beginTransaction();
+	Transaction tr = null;
 	try {
-	    session.delete(admin);
-	    tr.commit();
-	    if (log.isDebugEnabled()) {
-		log.debug("Delete Admin" + admin.toString());
+	    // check last modification
+	    if (admin.getLast_modified().equals(this.getAdmin(admin.getUser_name()).getLast_modified())) {
+		session = HibernateUtil.getSessionFactory().getCurrentSession();
+		tr = session.beginTransaction();
+		session.delete(admin);
+		tr.commit();
+		if (log.isDebugEnabled()) {
+		    log.debug("Delete Admin" + admin.toString());
+		}
+	    } else {
+		if (log.isDebugEnabled()) {
+		    log.debug("This is not the latest modification of admin " + admin.toString() + " so cannot delete");
+		}
+		// TODO:Create Exception
 	    }
 	} catch (RuntimeException ex) {
 	    tr.rollback();
@@ -119,17 +142,35 @@ public class AdminDAOImpl implements AdminDAO {
 
     @Override
     public void saveAdmin(Admin admin) {
-	session = HibernateUtil.getSessionFactory().getCurrentSession();
-	Transaction tr = session.beginTransaction();
-	try {// TODO:synchronize this
-	    session.save(admin);
-	    tr.commit();
-	    if (log.isDebugEnabled()) {
-		log.debug("Save Admin" + admin.toString());
+	Transaction tr = null;
+	try {
+	    // check whether the admin user name already exist
+	    if (this.isAdminUserNameExist(admin.getUser_name())) {
+		if (log.isDebugEnabled()) {
+		    log.debug("Admin username :" + admin.toString() + " already exist");
+		}
+		// TODO: Pass the Exception
+	    } else {
+		session = HibernateUtil.getSessionFactory().getCurrentSession();
+		tr = session.beginTransaction();
+		String dateTime = Formatters.formatDate(new Date());
+		admin.setDate_time(dateTime);
+		admin.setLast_modified(dateTime);
+		//set last modified data if scope is not null
+		if(admin.getAdminScopeSet()!=null){
+		    for(AdminScope adminScp:admin.getAdminScopeSet()){
+			adminScp.setLast_modified(dateTime);
+		    }
+		}
+		session.save(admin);
+		tr.commit();
+		if (log.isDebugEnabled()) {
+		    log.debug("Save Admin" + admin.toString());
+		}
 	    }
 	} catch (RuntimeException ex) {
-	    tr.rollback();
 	    log.error(ex);
+	    tr.rollback();
 	    // TODO: Throw exception
 	}
     }
@@ -170,8 +211,9 @@ public class AdminDAOImpl implements AdminDAO {
     @Override
     public List<Admin> getAllAdmins() {
 	session = HibernateUtil.getSessionFactory().getCurrentSession();
-	Transaction tr = session.beginTransaction();
+	Transaction tr = null;
 	try {
+	    tr = session.beginTransaction();
 	    List<Admin> adminAll = session.getNamedQuery(Admin.Constants.NAME_QUERY_FIND_ALL).list();
 	    tr.commit();
